% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomap.R
\name{geomap}
\alias{geomap}
\title{geomap}
\usage{
geomap(mapdata, series, match_on = NULL)
}
\arguments{
\item{mapdata}{Data frame of postcode values to be plotted}

\item{series}{The name of the data series to plot}

\item{match_on}{Optional. The name of the postcode level to join against. Must be one of:
\describe{
  \item{postcode_area}{e.g. SW}
  \item{postcode_district}{e.g. SE1}
  \item{postcode_region}{e.g. SE10}
  \item{postcode}{e.g. SE1 0SW}
}}
}
\value{
ggmap object
}
\description{
Creates a NUTS2-level ggmap from UK Postcode data
}
\examples{
geomap(sample_data, "population")
}
